pyproject.toml
[tool.black]
line-length = 120

[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[tomli]>=6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "lmms_engine"
version = "0.1.2"
authors = [{name = "LMMs-Lab"}]
description = "A simple, unified multimodal models training engine. Lean, flexible, and built for hacking, and hacking at scale."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "datasets",
    "hf_transfer",
    "transformers==4.57.1",
    "accelerate",
    "pillow",
    "peft",
    "deepspeed",
    "bitsandbytes",
    "triton",
    "loguru",
    "pydantic",
    "jsonlines",
    "librosa",
    "pynvml",
    "decord",
    "torch",
    "torchvision",
    "pyyaml",
    "wandb",
    "torchdata",
    "liger-kernel>=0.6.1",
    "qwen-vl-utils",
    "qwen-omni-utils",
    "einops",
    "pandas",
    "rich>=14.1.0",
    "hydra-core",
    "omegaconf",
    "requests",
    "tqdm",
    "opencv-python>=4.12.0.88",
]

[project.optional-dependencies]
pref = [
    "trl==0.16.1",
    "math-verify",
]

storage = [
    "google-cloud-storage",
    "azure-storage-blob",
    "azure-identity",
]

bagel = [
    "opencv_python",
]

sit = [
    "timm",
    "diffusers",
    "torchdiffeq",
]

all = [
    "lmms_engine[pref]",
    "lmms_engine[storage]",
]


[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = [
    "test"
]

[tool.wheel]
exclude = [
    "test"
]

[project.scripts]
lmms_launch="lmms_engine.launch.cli:main"


[project.urls]
Homepage = "https://lmms-lab.com/"
Repository = "https://github.com/EvolvingLMMs-Lab/lmms_engine"

[tool.isort]
known_third_party = ["wandb"]
